libcephfs: do not expose private symbols
otherwise the librados could be using the symbols leaked from libcephfs.
so a non-recursive lock could be still acquired when a thread is trying
to acquire it. actually it has already been relinquished by this thread.
because lockdep.cc was reoving the lock record from another instance of
`held`, and when it is acquired again, the `held` map still thinks that
the thread is holding the lock.

* CMakeLists.txt: hide all symbols from linked libs
* libcephfs.h: only expose public symbols

Signed-off-by: Kefu Chai <kchai@redhat.com>